(maint) General refactoring
- Grouping all files into correct folders
- Changing namespaces, etc
