+++ src/Chocolatey.Language.Server/Handlers/TextDocumentSyncHandler.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Schema;
+using Chocolatey.Language.Server.Engine;
 using Microsoft.Language.Xml;
 using OmniSharp.Extensions.Embedded.MediatR;
 using OmniSharp.Extensions.LanguageServer.Protocol;
@@ -15,7 +16,7 @@ using OmniSharp.Extensions.LanguageServer.Protocol.Server;
 using OmniSharp.Extensions.LanguageServer.Protocol.Server.Capabilities;
 using DiagnosticSeverity = OmniSharp.Extensions.LanguageServer.Protocol.Models.DiagnosticSeverity;
 
-namespace Chocolatey.Language.Server
+namespace Chocolatey.Language.Server.Handlers
 {
     public class TextDocumentSyncHandler : ITextDocumentSyncHandler
     {
