--- src/Chocolatey.Language.Server/ITypeLocator.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Reflection;
-
-namespace Chocolatey.Language.Server
-{
-    public interface ITypeLocator
-    {
-        IEnumerable<Type> GetTypesThatInheritOrImplement<T>();
-        IEnumerable<Type> GetTypesThatInheritOrImplement<T>(params Assembly[] Assemblies);
-    }
-}
